lesson_18/booking_app/requirements.txt
fastapi==0.110.2
uvicorn==0.29.0
pydantic_core==2.18.1
email_validator==2.1.1
dnspython==2.6.1